+++ demo/views/layout/layout.js
@@ -3,6 +3,7 @@ import RowDemo from './row-demo';
 import TabsDemo from './tabs-demo';
 import PodDemo from './pod-demo';
 import ContentDemo from './content-demo';
+import AppWrapperDemo from './app-wrapper-demo';
 import CarouselDemo from './carousel-demo';
 
 class Layout extends React.Component {
@@ -18,6 +19,7 @@ class Layout extends React.Component {
         <TabsDemo />
         <PodDemo />
         <PodDemo collapsed={ true } />
+        <AppWrapperDemo />
         <ContentDemo />
       </div>
     );